# GitRepoMind/server/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "GitRepoMind API",
                 "description":  "AI-powered GitHub Repository Analyzer \u0026 RAG Chat",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/chat":  {
                                    "post":  {
                                                 "tags":  [
                                                              "chat"
                                                          ],
                                                 "summary":  "Chat",
                                                 "description":  "Chat with the RAG-powered repository assistant.\n\nRetrieves relevant code chunks from the indexed repository,\ngenerates a RAG prompt, and uses an LLM to provide a conversational answer.\n\nArgs:\n    request: Chat request with repo_id, query, and optional session_id\n\nReturns:\n    Chat response with answer and sources\n\nRaises:\n    HTTPException: If repo not found or chat generation fails",
                                                 "operationId":  "chat_api_chat_post",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  "@{schema=}"
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  "@{application/json=}"
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  "@{application/json=}"
                                                                           }
                                                               }
                                             }
                                },
                  "/api/chat/history/{session_id}":  {
                                                         "get":  {
                                                                     "tags":  [
                                                                                  "chat"
                                                                              ],
                                                                     "summary":  "Get Chat History",
                                                                     "description":  "Retrieve conversation history for a session.\n\nArgs:\n    session_id: Conversation session ID\n\nReturns:\n    HistoryResponse with list of messages",
                                                                     "operationId":  "get_chat_history_api_chat_history__session_id__get",
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "session_id",
                                                                                            "in":  "path",
                                                                                            "required":  true,
                                                                                            "schema":  "@{type=string; title=Session Id}"
                                                                                        }
                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "Successful Response",
                                                                                                   "content":  "@{application/json=}"
                                                                                               },
                                                                                       "422":  {
                                                                                                   "description":  "Validation Error",
                                                                                                   "content":  "@{application/json=}"
                                                                                               }
                                                                                   }
                                                                 },
                                                         "delete":  {
                                                                        "tags":  [
                                                                                     "chat"
                                                                                 ],
                                                                        "summary":  "Delete Chat History",
                                                                        "description":  "Delete conversation history for a session.\n\nArgs:\n    session_id: Conversation session ID\n\nReturns:\n    Empty response on success\n\nRaises:\n    HTTPException: If operation fails",
                                                                        "operationId":  "delete_chat_history_api_chat_history__session_id__delete",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "session_id",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  "@{type=string; title=Session Id}"
                                                                                           }
                                                                                       ],
                                                                        "responses":  {
                                                                                          "204":  {
                                                                                                      "description":  "Successful Response"
                                                                                                  },
                                                                                          "422":  {
                                                                                                      "description":  "Validation Error",
                                                                                                      "content":  "@{application/json=}"
                                                                                                  }
                                                                                      }
                                                                    }
                                                     },
...
              },
...
}